Add push command
This is intended to be a simpler `git-review` script with more power.
One of the major selling points is that one does not need to specify an
explicit list of users--the command will query the Gerrit server and
return all relevant users/groups matching a given regex.

Thus this makes it really easy to tag multiple people for review.

This command also provides helpers for specifying whether a review is a
draft or which topic to associate with the pushed commits.