Add option for command-line reviewer count opt-in
I don't know if you appreciate how much I like not having to interact
with my `gerrit push` command :). This adds an option so I can set a
higher count of people that I'm okay with spamming, in the (very real)
case that I have a team size right on the threshold that the gem sets.